rename RQB folder
