+++ src/components/filter/expert/expert-filter-form.tsx
@@ -26,7 +26,7 @@ import { FieldType } from './expert-filter.type';
 import { v4 as uuid4 } from 'uuid';
 import { useIntl } from 'react-intl';
 import { FieldConstants } from '../../../utils/field-constants';
-import CustomReactQueryBuilder from '../../react-query-builder-inputs/custom-react-query-builder';
+import CustomReactQueryBuilder from '../../react-query-builder/custom-react-query-builder';
 import InputWithPopupConfirmation from '../../inputs/react-hook-form/select-inputs/input-with-popup-confirmation';
 import SelectInput from '../../inputs/react-hook-form/select-inputs/select-input';
 import { FilterType } from '../constants/filter-constants';
